[IMP] mail: keep composer actions on same line
Before this commit, when typing a long message in the composer and
the textarea grows, the buttons were moving as to be centered with
the textarea.

This didn't look good, because usually buttons are expected to stay
in place except when they are visually inline with some other
content, but that's not the case here.

closes odoo/odoo#169521

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>